requirements.txt
requests
python-dotenv
openai
tqdm
